requirements.txt
django-background-tasks==1.2.0
django-compat==1.0.15
numpy==1.16.5
Pillow==6.1.0
six==1.12.0
psycopg2==2.6.2